+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Menus/VertexColorRemoverMenu.cs
@@ -3,7 +3,6 @@
 // Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
 // </copyright>
 
-#if VQT_HAS_VRCSDK_BASE
 using KRT.VRCQuestTools.Components;
 using UnityEditor;
 using UnityEngine;
@@ -34,7 +33,7 @@ namespace KRT.VRCQuestTools.Menus
             }
             remover.includeChildren = true;
             remover.enabled = true;
-            Debug.Log($"[{VRCQuestTools.Name}] All vertex colors will be removed from {target}", target);
+            Logger.Log($"All vertex colors will be removed from {target}", target);
         }
 
         [MenuItem(VRCQuestToolsMenus.MenuPaths.RemoveAllVertexColors, true)]
@@ -45,4 +44,3 @@ namespace KRT.VRCQuestTools.Menus
         }
     }
 }
-#endif