+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/NDMF/Errors/MissingNetworkIDAssignerWarning.cs
@@ -1,9 +1,5 @@
 ﻿using nadena.dev.ndmf;
-#if VQT_HAS_NDMF_ERROR_REPORT
 using nadena.dev.ndmf.localization;
-#else
-using KRT.VRCQuestTools.Ndmf.Dummy;
-#endif
 
 namespace KRT.VRCQuestTools.Ndmf
 {