+++ Packages/com.github.kurotu.vrc-quest-tools/Runtime/Components/Interfaces/IMaterialConversionComponent.cs
@@ -1,3 +1,4 @@
+using System.Linq;
 using KRT.VRCQuestTools.Models;
 
 namespace KRT.VRCQuestTools.Components
@@ -31,5 +32,16 @@ namespace KRT.VRCQuestTools.Components
         /// Gets a value indicating whether to use default material convert settings.
         /// </summary>
         bool IsPrimaryRoot { get; }
+
+        /// <summary>
+        /// Gets the cache key for the material conversion component.
+        /// </summary>
+        /// <returns>Cache key.</returns>
+        public string GetCacheKey()
+        {
+            return $"{DefaultMaterialConvertSettings.GetCacheKey()}_" +
+                   $"{string.Join("_", AdditionalMaterialConvertSettings.Select(s => s.GetCacheKey()))}_" +
+                   $"{RemoveExtraMaterialSlots}";
+        }
     }
 }